Module 5

+++ app/controllers/answers_controller.rb
@@ -1,20 +1,17 @@
 # frozen_string_literal: true
 
 class AnswersController < ApplicationController
-  def show; end
-
-  def new; end
-
-  def edit; end
+  before_action :authenticate_user!, except: %i[show]
+  before_action :load_question, only: %i[new create]
 
   def create
-    @question = Question.find(params[:question_id])
     @answer = @question.answers.new(answer_params)
+    current_user.answers.push(@answer)
 
     if @answer.save
-      redirect_to @answer
+      redirect_to question_path(@question), notice: 'Your answer successfully created'
     else
-      render :new
+      render :'questions/show'
     end
   end
 
@@ -27,11 +24,19 @@ class AnswersController < ApplicationController
   end
 
   def destroy
-    answer.destroy
-    redirect_to question_path(answer.question)
+    if current_user.author_of?(answer)
+      answer.destroy
+      redirect_to question_path(answer.question), notice: 'Answer was deleted successful'
+    else
+      redirect_to question_path(answer.question), alert: "You don't have permission"
+    end
   end
 
   private
+
+  def load_question
+    @question = Question.find(params[:question_id])
+  end
 
   def answer
     @answer ||= params[:id] ? Answer.find(params[:id]) : Answer.new
